+++ apps/explorer/lib/explorer/chain/log.ex
@@ -35,10 +35,10 @@ defmodule Explorer.Chain.Log do
           block_hash: Hash.Full.t(),
           block_number: non_neg_integer() | nil,
           data: Data.t(),
-          first_topic: String.t(),
-          second_topic: String.t(),
-          third_topic: String.t(),
-          fourth_topic: String.t(),
+          first_topic: Hash.Full.t(),
+          second_topic: Hash.Full.t(),
+          third_topic: Hash.Full.t(),
+          fourth_topic: Hash.Full.t(),
           transaction: %Ecto.Association.NotLoaded{} | Transaction.t(),
           transaction_hash: Hash.Full.t(),
           index: non_neg_integer(),
@@ -48,10 +48,10 @@ defmodule Explorer.Chain.Log do
   @primary_key false
   schema "logs" do
     field(:data, Data)
-    field(:first_topic, :string)
-    field(:second_topic, :string)
-    field(:third_topic, :string)
-    field(:fourth_topic, :string)
+    field(:first_topic, Hash.Full)
+    field(:second_topic, Hash.Full)
+    field(:third_topic, Hash.Full)
+    field(:fourth_topic, Hash.Full)
     field(:index, :integer, primary_key: true)
     field(:type, :string)
     field(:block_number, :integer)
@@ -190,7 +190,8 @@ defmodule Explorer.Chain.Log do
   end
 
   defp find_method_candidates(log, transaction, options, events_acc, skip_sig_provider?) do
-    with "0x" <> hex_part <- log.first_topic,
+    with false <- is_nil(log.first_topic),
+         hex_part <- Base.encode16(log.first_topic.bytes, case: :lower),
          {number, ""} <- Integer.parse(hex_part, 16) do
       <<method_id::binary-size(4), _rest::binary>> = :binary.encode_unsigned(number)
 
@@ -201,7 +202,8 @@ defmodule Explorer.Chain.Log do
         {result, Map.put(events_acc, method_id, result)}
       end
     else
-      _ -> {{:error, :could_not_decode}, events_acc}
+      _ ->
+        {{:error, :could_not_decode}, events_acc}
     end
   end
 
@@ -243,10 +245,10 @@ defmodule Explorer.Chain.Log do
            abi
            |> ABI.parse_specification(include_events?: true)
            |> Event.find_and_decode(
-             decode16!(log.first_topic),
-             decode16!(log.second_topic),
-             decode16!(log.third_topic),
-             decode16!(log.fourth_topic),
+             log.first_topic && log.first_topic.bytes,
+             log.second_topic && log.second_topic.bytes,
+             log.third_topic && log.third_topic.bytes,
+             log.fourth_topic && log.fourth_topic.bytes,
              log.data.bytes
            ) do
       {:ok, selector, mapping}